Use redirect pattern for transaction->internal transaction list
