+++ apps/explorer_web/lib/explorer_web/views/transaction_internal_transaction_view.ex
@@ -1,0 +1,9 @@
+defmodule ExplorerWeb.TransactionInternalTransactionView do
+  use ExplorerWeb, :view
+  @dialyzer :no_match
+
+  alias ExplorerWeb.TransactionView
+
+  defdelegate value(txn, opts), to: TransactionView
+  defdelegate gas(txn), to: TransactionView
+end
