+++ apps/explorer_web/test/support/conn_case.ex
@@ -25,6 +25,7 @@ defmodule ExplorerWeb.ConnCase do
       @endpoint ExplorerWeb.Endpoint
 
       import Explorer.Factory
+      import ExplorerWeb.Factory
     end
   end
